i386/acpi: restore device paths for 6.2 vms
After fixing the _UID value for the primary PCI root bridge in
af1b80ae it was discovered that this change updates Windows
configuration in an incompatible way causing network configuration
failure unless DHCP is used.
This change reverts the _UID update from 1 to 0 for i440fx
VMs before version 6.2 to maintain the original behaviour when
upgrading.

[sc-73907]